Seperated First into 3 other components, Login/Register/UploadPokemon.jsx
